Move stmt_from_rule to reporting
